# src/docs/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "CarKit API Documentation",
    "version": "0.0.1"
  },
  "host": [
    "carkit.lazyy.fr"
  ],
  "schemes": [
    "https"
  ],
  "paths": {
    "/user/me": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Get all information of the connected user.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "desc": "fullUser"
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          }
        }
      }
    },
    "/user/": {
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Delete an User.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {
              "status": 200,
              "msg": "User bien supprimé"
            }
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Unauthorized."
            }
          }
        }
      }
    },
    "/user/register": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "Register a new user.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "User registration data",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "user@example.com"
                },
                "password": {
                  "example": "Password123"
                },
                "image": {
                  "example": "base64ImageString"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {}
          }
        }
      }
    },
    "/user/update": {
      "put": {
        "tags": [
          "User"
        ],
        "description": "Update user details.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "User update data",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "new@example.com"
                },
                "firstName": {
                  "example": "John"
                },
                "lastName": {
                  "example": "Doe"
                },
                "password": {
                  "example": "newPass"
                },
                "username": {
                  "example": "johndoe"
                },
                "image": {
                  "example": "base64ImageString"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {}
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "User not found."
            }
          },
          "422": {
            "description": "Response",
            "examples": {
              "status": 422,
              "msg": "Required fields missing."
            }
          }
        }
      }
    },
    "/user/image": {
      "delete": {
        "tags": [
          "User"
        ],
        "description": "Delete user image.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {}
          },
          "404": {
            "description": "Response",
            "examples": {
              "status": 404,
              "msg": "User not found."
            }
          },
          "422": {
            "description": "Response",
            "examples": {
              "status": 422,
              "msg": "Required fields missing."
            }
          }
        }
      }
    },
    "/user/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "User login.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "description": "User login data",
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "example": "user@example.com"
                },
                "password": {
                  "example": "Password123"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {}
          }
        }
      }
    },
    "/user/refresh-token/{refreshToken}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "Refresh user token.",
        "parameters": [
          {
            "in": "path",
            "name": "refreshToken",
            "required": true,
            "description": "Path parameter refreshToken",
            "schema": {
              "type": "string",
              "example": "yourRefreshToken"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response",
            "examples": {}
          },
          "401": {
            "description": "Response",
            "examples": {
              "status": 401,
              "msg": "Aucun token valide trouvé."
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your Bearer token in the format **Bearer <token>**"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
...
}